requirements.txt
django
djangorestframework
environs
Pillow
psycopg2-binary
tensorflow
pillow
numpy
matplotlib
openai